requirements.txt
streamlit>=1.40.0
groq>=0.11.0
# Chroma <0.5.20 arrastra opentelemetry-otlp-grpc con _pb2.py viejo que
# rompe con protobuf>=4.21 en Streamlit Cloud. Subir la cota.
chromadb>=0.5.20
tiktoken>=0.7.0
python-dotenv>=1.0.0
